fix: refactor and reuse code
