+++ src/containers/Auth/Auth.js
@@ -6,6 +6,7 @@ import classes from './Auth.module.css';
 import * as actions from '../../store/actions';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
+import { updateObject, checkValidity } from '../../shared/utility';
 
 class Auth extends Component {
     state = {
@@ -44,48 +45,16 @@ class Auth extends Component {
 
     }
 
-    checkValidity(value, rules) {
-        if (!rules) {
-            return true
-        }
-
-        let isValid = true;
-
-        if (rules.requried) {
-            isValid = value.trim() !== '' && isValid;
-        }
-
-        if (rules.minLength) {
-            isValid = value.length >= rules.minLength && isValid;
-        }
-
-        if (rules.maxLength) {
-            isValid = value.length <= rules.maxLength && isValid;
-        }
-
-        if (rules.isEmail) {
-            const pattern = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;
-            isValid = pattern.test(value) && isValid;
-        }
-
-        if (rules.isNumeric) {
-            const pattern = /^\d+$/;
-            isValid = pattern.test(value) && isValid;
-        }
-
-        return isValid;
-    }
+  
 
     inputChangedHandler(event, element) {
-        const updatedControls = {
-            ...this.state.controls,
-            [element]: {
-                ...this.state.controls[element],
+        const updatedControls = updateObject(this.state.controls, {
+            [element]: updateObject(this.state.controls[element], {
                 value: event.target.value,
-                valid: this.checkValidity(event.target.value, this.state.controls[element].validation),
+                valid: checkValidity(event.target.value, this.state.controls[element].validation),
                 touched: true
-            }
-        }
+            })
+        }); 
 
         this.setState({ controls: updatedControls });
     }
