+++ src/containers/Checkout/ContactData/ContactData.module.css
@@ -7,9 +7,4 @@
 }
 
 
-@media (min-width: 600px) {
-    .ContactData {
-        width: 500px;
-    }
-}
