+++ src/shared/utility.js
@@ -1,0 +1,38 @@
+export const updateObject = (oldObject, updatedProps) => {
+    return {
+        ...oldObject,
+        ...updatedProps
+    }
+}
+
+export const checkValidity = (value, rules) => {
+    if (!rules) {
+        return true
+    }
+
+    let isValid = true;
+
+    if (rules.requried) {
+        isValid = value.trim() !== '' && isValid;
+    }
+
+    if (rules.minLength) {
+        isValid = value.length >= rules.minLength && isValid;
+    }
+
+    if (rules.maxLength) {
+        isValid = value.length <= rules.maxLength && isValid;
+    }
+
+    if (rules.isEmail) {
+        const pattern = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;
+        isValid = pattern.test(value) && isValid;
+    }
+
+    if (rules.isNumeric) {
+        const pattern = /^\d+$/;
+        isValid = pattern.test(value) && isValid;
+    }
+
+    return isValid;
+}
