--- src/store/utility.js
@@ -1,6 +1,0 @@
-export const updateObject = (oldObject, updatedProps) => {
-    return {
-        ...oldObject,
-        ...updatedProps
-    }
-}
